Handle error response

+++ domain/src/main/java/com/nuweather/domain/model/CurrentWeather.kt
@@ -1,30 +1,30 @@
 package com.nuweather.domain.model
 
 data class CurrentWeather(
-    val id: Int,
+    val id: Long,
     val description: String,
     val temp: Double,
     val tempMin: Double,
     val tempMax: Double,
-    val pressure: Int,
-    val visibility: Int,
-    val clouds: Int,
+    val pressure: Double,
+    val visibility: Double,
+    val clouds: Double,
     val wind: Wind,
     val sys: Sys,
-    val dt: Int
+    val dt: Long
 ) : Model()
 
 data class Wind(
     val speed: Double,
-    val deg: Int
+    val deg: Double
 ) : Model()
 
 
 data class Sys(
     val type: Int,
-    val id: Int,
+    val id: Long,
     val message: Double,
     val country: String,
-    val sunrise: Int,
-    val sunset: Int
+    val sunrise: Long,
+    val sunset: Long
 ) : Model()
